requirements.txt
PyJWT==1.7.1
requests==2.24.0
Flask==1.1.2
urllib3==1.25.11